Fix web demo build
For some reason, after the clap feature change in #129 the build of the
web demo is broken. What's even more strange is that it only fails at
build time (not at type-check time). Maybe it's because of some build
script that no longer compiles?

In any case, upgrading both `clap` and `windows-sys` fixes the problem.

+++ front-end/tests/utils.rs
@@ -35,14 +35,7 @@ pub mod windows {
         let handle = std::fs::OpenOptions::new().write(true).open(path)?;
         unsafe {
             let mut overlapped: OVERLAPPED = std::mem::zeroed();
-            let ret = LockFileEx(
-                handle.as_raw_handle() as isize,
-                flags,
-                0,
-                !0,
-                !0,
-                &mut overlapped,
-            );
+            let ret = LockFileEx(handle.as_raw_handle(), flags, 0, !0, !0, &mut overlapped);
 
             if ret == 0 {
                 return Err(Box::new(std::io::Error::last_os_error()));
